#8.14 Bugfix

+++ src/controllers/videoController.js
@@ -8,17 +8,24 @@ export const home = async (req, res) => {
 export const watch = async (req, res) => {
   const { id } = req.params;
   const video = await Video.findById(id).populate("owner");
-  console.log(video);
   if (video === null) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
   }
   return res.render("videos/watch", { pageTitle: video.title, video });
 };
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const {
+    session: {
+      user: { _id },
+    },
+    params: { id },
+  } = req;
   const video = await Video.findById(id);
   if (video === null) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("videos/edit", {
     pageTitle: `Edit: ${video.title}`,
@@ -26,11 +33,19 @@ export const getEdit = async (req, res) => {
   });
 };
 export const postEdit = async (req, res) => {
-  const { id } = req.params;
-  const { title, description, hashtags } = req.body;
-  const video = await Video.exists({ _id: id });
+  const {
+    session: {
+      user: { _id },
+    },
+    params: { id },
+    body: { title, description, hashtags },
+  } = req;
+  const video = await Video.findById({ _id: id });
   if (!video) {
     return res.render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -74,7 +89,22 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
-  const { id } = req.params;
+  const {
+    session: {
+      user: { _id },
+    },
+    params: { id },
+  } = req;
+  const video = await Video.findById(id).populate("owner");
+  const { _id: ownerId, videos } = video.owner;
+  if (!video) {
+    return res.render("404", { pageTitle: "Video not found." });
+  }
+  if (String(ownerId) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  videos.splice(videos.indexOf(id), 1);
+  await User.findByIdAndUpdate(ownerId, { videos });
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
